+++ src/pages/Shuttle/ClaimModal/ShuttleButton/Out.js
@@ -2,17 +2,13 @@ import {useState, useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {useTranslation} from 'react-i18next'
 import Big from 'big.js'
-import {MaxUint256} from '@ethersproject/constants'
 
-import {Button, Loading} from '../../../../components'
+import {Button} from '../../../../components'
 import {SupportedChains, KeyOfCfx} from '../../../../constants/chainConfig'
 import useShuttleAddress from '../../../../hooks/useShuttleAddress'
 import {useIsCfxChain, useIsBtcChain} from '../../../../hooks'
 import {useShuttleContract} from '../../../../hooks/useShuttleContract'
-import {
-  ContractType,
-  ContractConfig,
-} from '../../../../constants/contractConfig'
+import {ContractType} from '../../../../constants/contractConfig'
 import {useCustodianData} from '../../../../hooks/useShuttleData'
 import {
   ZeroAddrHex,
@@ -23,8 +19,6 @@ import {
 import {useShuttleState} from '../../../../state'
 import {getExponent, calculateGasMargin, updateTx} from '../../../../utils'
 import {useTxState} from '../../../../state/transaction'
-import {useTokenContract, useTokenAllowance} from '../../../../hooks/usePortal'
-import {useIsNativeToken} from '../../../../hooks/useWallet'
 import {requestUserOperationByHash} from '../../../../utils/api'
 
 function ShuttleOutButton({
@@ -41,7 +35,6 @@ function ShuttleOutButton({
   setSendStatus,
 }) {
   const {t} = useTranslation()
-  const {display_symbol} = fromToken
   const {origin, decimals, ctoken} = toToken
   const isCfxChain = useIsCfxChain(origin)
   const isToChainBtc = useIsBtcChain(toChain)
@@ -62,17 +55,6 @@ function ShuttleOutButton({
   const [didMount, setDidMount] = useState(false)
   const {unshiftTx, transactions, setTransactions} = useTxState()
   window._transactions = new Map(Object.entries(transactions))
-  const drContractAddress =
-    ContractConfig[ContractType.depositRelayerCfx]?.address?.[toChain]
-  const [approveShown, setApproveShown] = useState(false)
-  const [isApproving, setIsApproving] = useState(false)
-  const [fetchApprove, setFetchApprove] = useState(false)
-  const tokenContract = useTokenContract(ctoken)
-  const tokenAllownace = useTokenAllowance(ctoken, [
-    fromAddress,
-    drContractAddress,
-  ])
-  const isNativeToken = useIsNativeToken(fromChain, ctoken)
   useEffect(() => {
     setDidMount(true)
     if (isToChainBtc) {
@@ -84,24 +66,6 @@ function ShuttleOutButton({
       setDidMount(false)
     }
   }, [isToChainBtc, toAddress, toBtcAddress])
-
-  useEffect(() => {
-    setDidMount(true)
-    if (
-      isCfxChain &&
-      !isNativeToken &&
-      new Big(tokenAllownace.toString(10)).lt(
-        new Big(value).times(getExponent(decimals)),
-      )
-    ) {
-      setApproveShown(true)
-    } else {
-      setApproveShown(false)
-    }
-    return () => {
-      setDidMount(false)
-    }
-  }, [decimals, tokenAllownace, value, isNativeToken, isCfxChain, fetchApprove])
 
   function getShuttleStatusData(hash, type = TypeTransaction.transaction) {
     let fee = out_fee ? out_fee.toString(10) : '0'
@@ -221,75 +185,12 @@ function ShuttleOutButton({
     }
   }
 
-  function contractApprove(tokenContract, value, gas, storage) {
-    tokenContract
-      .approve(drContractAddress, value)
-      .sendTransaction({
-        gas: gas ? calculateGasMargin(gas, 0.5) : undefined,
-        from: fromAddress,
-        storageLimit: storage ? calculateGasMargin(storage, 0.5) : undefined,
-      })
-      .confirmed()
-      .then(receipt => {
-        unshiftTx(
-          getShuttleStatusData(
-            receipt?.transactionHash,
-            TypeTransaction.approve,
-          ),
-        )
-        setFetchApprove(!fetchApprove)
-        setIsApproving(false)
-        setApproveShown(false)
-      })
-      .catch(() => {
-        setIsApproving(false)
-      })
-  }
-
-  const onApprove = async () => {
-    if (isApproving) return
-    setIsApproving(true)
-    //MaxUint256
-    tokenContract
-      .approve(drContractAddress, MaxUint256)
-      .estimateGasAndCollateral({
-        from: fromAddress,
-      })
-      .then(estimateData => {
-        contractApprove(
-          tokenContract,
-          MaxUint256,
-          estimateData?.gasLimit,
-          estimateData?.storage,
-        )
-      })
-      .catch(error => {
-        if (error.data && error.data.code === -32000) {
-          contractApprove(tokenContract, 0)
-        } else {
-          setIsApproving(false)
-        }
-      })
-  }
-
   if (!didMount) {
     return null
   }
   return (
     <>
-      {approveShown && (
-        <Button
-          fullWidth
-          onClick={onApprove}
-          disabled={disabled}
-          size="large"
-          id="approve"
-        >
-          {isApproving && <Loading className="!w-6 !h-6" />}
-          {!isApproving && t('approve', {tokenSymbol: display_symbol})}
-        </Button>
-      )}
-      {!approveShown && (
+      {
         <Button
           onClick={onSubmit}
           disabled={disabled}
@@ -298,7 +199,7 @@ function ShuttleOutButton({
         >
           {t('send')}
         </Button>
-      )}
+      }
     </>
   )
 }
